requirements.txt
torch
torchvision
pandas
streamlit
Pillow